Working to reduce the memory footprint of Trace - Characterization
Testing the changes to the characterization meant to reduce the memory footprint. Doing this for the large alignment takes a long time so I will continue working while this test is running.

+++ src/SupportingClasses/Trace.py
@@ -248,8 +248,9 @@ def save_freq_table(freq_table, low_memory, node_name, out_dir):
     """
     if low_memory:
         fn = os.path.join(out_dir, '{}_freq_table.pkl'.format(node_name))
-        with open(fn, 'wb') as handle:
-            pickle.dump(freq_table, handle, pickle.HIGHEST_PROTOCOL)
+        if not os.path.isfile(fn):
+            with open(fn, 'wb') as handle:
+                pickle.dump(freq_table, handle, pickle.HIGHEST_PROTOCOL)
         freq_table = fn
     return freq_table
 
@@ -349,11 +350,13 @@ def characterization(processor):
         if single:
             pos_table.compute_frequencies()
             pos_table.to_csv(os.path.join(u_dir, '{}_position_freq_table.tsv'.format(node.name)))
-            pos_table = save_freq_table(freq_table=pos_table, low_memory=low_mem, node_name=node.name, out_dir=u_dir)
+            pos_table = save_freq_table(freq_table=pos_table, low_memory=low_mem,
+                                        node_name='{}_single'.format(node.name), out_dir=u_dir)
         if pair:
             pair_table.compute_frequencies()
             pair_table.to_csv(os.path.join(u_dir, '{}_pair_freq_table.tsv'.format(node.name)))
-            pair_table = save_freq_table(freq_table=pair_table, low_memory=low_mem, node_name=node.name, out_dir=u_dir)
+            pair_table = save_freq_table(freq_table=pair_table, low_memory=low_mem,
+                                         node_name='{}_pair'.format(node.name), out_dir=u_dir)
         # Store the current nodes characterization in the shared dictionary.
         tables = {'single': pos_table, 'pair': pair_table}
         freq_tables[node.name] = tables
